+++ dallinger/docker/tools.py
@@ -24,7 +24,7 @@ docker_compose_template = Template(
 
 
 class DockerComposeWrapper:
-    """Wrapper around a docker compose local daemon, modeled after HerokuLocalWrapper.
+    """Wrapper around a docker compose local daemon, modeled after LocalProcfileWrapper.
 
     Provides for verified startup and shutdown, and allows observers to register
     to recieve subprocess output via 'monitor()'.