+++ tests/test_deployment.py
@@ -1933,10 +1933,10 @@ class TestDebugServer:
 
     @pytest.fixture
     def debugger(self, debugger_unpatched):
-        from dallinger.heroku.tools import HerokuLocalWrapper
+        from dallinger.heroku.tools import LocalProcfileWrapper
 
         debugger = debugger_unpatched
-        debugger.notify = mock.Mock(return_value=HerokuLocalWrapper.MONITOR_STOP)
+        debugger.notify = mock.Mock(return_value=LocalProcfileWrapper.MONITOR_STOP)
         return debugger
 
     def test_startup(self, debugger):
@@ -2018,7 +2018,7 @@ class TestDebugServer:
             p.expect_exact("Server is running", timeout=300)
             p.expect_exact("Recruitment is complete", timeout=600)
             p.expect_exact("Experiment completed", timeout=60)
-            p.expect_exact("Local Heroku process terminated", timeout=10)
+            p.expect_exact("Local server processes terminated", timeout=10)
         finally:
             try:
                 p.sendcontrol("c")
@@ -2143,10 +2143,10 @@ class TestLoad:
     @pytest.fixture
     def loader(self, db_session, output, clear_workers):
         from dallinger.deployment import LoaderDeployment
-        from dallinger.heroku.tools import HerokuLocalWrapper
+        from dallinger.heroku.tools import LocalProcfileWrapper
 
         loader = LoaderDeployment(self.exp_id, output, verbose=True, exp_config={})
-        loader.notify = mock.Mock(return_value=HerokuLocalWrapper.MONITOR_STOP)
+        loader.notify = mock.Mock(return_value=LocalProcfileWrapper.MONITOR_STOP)
 
         yield loader
 
@@ -2160,10 +2160,10 @@ class TestLoad:
         loader.keep_running = mock.Mock(return_value=False)
 
         def launch_and_finish(self):
-            from dallinger.heroku.tools import HerokuLocalWrapper
+            from dallinger.heroku.tools import LocalProcfileWrapper
 
             loader.out.log("Launching replay browser...")
-            return HerokuLocalWrapper.MONITOR_STOP
+            return LocalProcfileWrapper.MONITOR_STOP
 
         loader.start_replay = mock.Mock(
             return_value=None, side_effect=launch_and_finish
@@ -2176,7 +2176,7 @@ class TestLoad:
 
         loader.out.log.assert_has_calls(
             [
-                mock.call("Starting up the Heroku Local server..."),
+                mock.call("Starting up the honcho Local server..."),
                 mock.call("Ingesting dataset from some_experiment_id-data.zip..."),
                 mock.call(
                     "Server is running on http://localhost:{}. Press Ctrl+C to exit.".format(
@@ -2184,8 +2184,8 @@ class TestLoad:
                     )
                 ),
                 mock.call("Terminating dataset load for experiment some_experiment_id"),
-                mock.call("Cleaning up local Heroku process..."),
-                mock.call("Local Heroku process terminated."),
+                mock.call("Cleaning up local server processes..."),
+                mock.call("Local server processes terminated."),
             ]
         )
 
@@ -2200,7 +2200,7 @@ class TestLoad:
 
         replay_loader.out.log.assert_has_calls(
             [
-                mock.call("Starting up the Heroku Local server..."),
+                mock.call("Starting up the honcho Local server..."),
                 mock.call("Ingesting dataset from some_experiment_id-data.zip..."),
                 mock.call(
                     "Server is running on http://localhost:{}. Press Ctrl+C to exit.".format(
@@ -2210,7 +2210,7 @@ class TestLoad:
                 mock.call("Launching the experiment..."),
                 mock.call("Launching replay browser..."),
                 mock.call("Terminating dataset load for experiment some_experiment_id"),
-                mock.call("Cleaning up local Heroku process..."),
-                mock.call("Local Heroku process terminated."),
+                mock.call("Cleaning up local server processes..."),
+                mock.call("Local server processes terminated."),
             ]
         )